# Faster and cached square root