--- a/math/math_helper.go
+++ b/math/math_helper.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	stdmath "math"
 	"math/bits"
+	"sync"
 
 	"github.com/thomaso-mirodin/intmath/u64"
 )
@@ -26,6 +27,12 @@ var (
 	ErrMulOverflow  = errors.New("multiplication overflows")
 	ErrAddOverflow  = errors.New("addition overflows")
 	ErrSubUnderflow = errors.New("subtraction underflow")
+
+	// Sensible guess for 500 000 validators
+	cachedSquareRoot = struct {
+		sync.Mutex
+		squareRoot, balance uint64
+	}{squareRoot: 126491106, balance: 15999999897103236}
 )
 
 // Common square root values.
@@ -41,6 +48,25 @@ var squareRootTable = map[uint64]uint64{
 	262144:  512,
 	1048576: 1024,
 	4194304: 2048,
+}
+
+// SquareRootEffectiveBalance implements Newton's algorithm to compute the square root of
+// the given uint64 starting from the last cached value
+func SquareRootEffectiveBalance(balance uint64) uint64 {
+	cachedSquareRoot.Lock()
+	defer cachedSquareRoot.Unlock()
+	if balance == cachedSquareRoot.balance {
+		return cachedSquareRoot.squareRoot
+	}
+	cachedSquareRoot.balance = balance
+	val := balance / cachedSquareRoot.squareRoot
+	for {
+		cachedSquareRoot.squareRoot = (cachedSquareRoot.squareRoot + val) / 2
+		val = balance / cachedSquareRoot.squareRoot
+		if cachedSquareRoot.squareRoot <= val {
+			return cachedSquareRoot.squareRoot
+		}
+	}
 }
 
 // IntegerSquareRoot defines a function that returns the