@potuz 2023-03-29
Tested that the function returns the same value as our previous method for balances between 700K validators and 5 million validators.